+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 - **Working Settings**
   - Enable Scheduler UI
+  - All Custom Scan Options
 - System Tray
 ### Improved
 - The `useSettings` hook